\name{bsp1}
\alias{bsp1}
\docType{data}
\title{ Datatable of Example 1}
\description{
  This is a data-table, used in example 1
}
\usage{data(bsp1)}
\format{
  A data frame with 10 observations on the following 4 variables.
  \describe{
    \item{\code{Geschlecht}}{a factor with levels \code{m} \code{w}, giving the sex}
    \item{\code{Alter}}{a numeric vector}
    \item{\code{Gewicht}}{a numeric vector, giving the weight}
    \item{\code{Groesse}}{a numeric vector, giving the height}
  }
}

\source{
  \url{https://de.wikibooks.org/wiki/GNU_R:_Anwendungsbeispiele#Beispiel_1}
}

\keyword{datasets}
