\name{bsp2}
\alias{bsp2}
\docType{data}
\title{ Datatable of Example 2}
\description{
  This is a data-table, used in example 2
}
\usage{data(bsp2)}
\format{
  A data frame with 20 observations on the following 2 variables.
  \describe{
    \item{\code{Geschlecht}}{a factor with levels \code{m} \code{w}, giving the sex}
    \item{\code{Note}}{a numeric vector, giving the grade}
  }
}

\source{
  \url{http://de.wikibooks.org/wiki/GNU_R:_Anwendungsbeispiele#Beispiel_2}
}
\keyword{datasets}
