\name{bsp3}
\alias{bsp3}
\docType{data}
\title{ Datatable of Example 3}
\description{
 This is a data-table, used in example 3
}
\usage{data(bsp3)}
\format{
  A data frame with 14 observations on the following 2 variables.
  \describe{
    \item{\code{Erfolg}}{a factor with levels \code{0} \code{1}, giving success}
    \item{\code{Abschlussnote}}{a numeric vector, giving the grade}
  }
}

\source{
  \url{https://de.wikibooks.org/wiki/GNU_R:_Anwendungsbeispiele#Beispiel_3}
}
\keyword{datasets}
