\name{bsp4}
\alias{bsp4}
\docType{data}
\title{ Datatable of Example 4}
\description{
 This is a data-table, used in example 4
}
\usage{data(bsp4)}
\format{
  A data frame with 7 observations on the following 4 variables.
  \describe{
    \item{\code{Name}}{a factor  giving people's names}
    \item{\code{Geschlecht}}{a factor giving people's sex}
    \item{\code{Lieblingsfarbe}}{a factor giving people's colour}
    \item{\code{Einkommen}}{a numeric vector, giving people's earnings}
  }
}

\source{
  \url{http://de.wikibooks.org/wiki/GNU_R:_Anwendungsbeispiele#Beispiel_4}
}

\keyword{datasets}
