\name{cms}
\alias{cms}
\docType{data}
\title{ Dataset of an assessment instrument}
\description{
  A dataset ment to be used for the "sens.spec"-function of the wikibooks-package
}
\usage{data(cms)}
\format{
  A data frame with 620 observations on the following 2 variables.
  \describe{
    \item{\code{ascore}}{scores of the assessment-instrument}
    \item{\code{arisk}}{definition whether "ascore" leads to positive group (1) oder negative group (0)}
  }
}
\details{
 This dataset contains the scores of an assessment-instrument and whether the probands belong to positive or negative group.
 Data is ment to be used with the "sens.spec"-function of the wikibook-package.
}
\source{
 \url{https://de.wikibooks.org/wiki/GNU_R:_Programmierbeispiele#Beispiel_2}
}
\keyword{datasets}
