\name{Essen.Zeit}
\alias{Essen.Zeit}
\docType{data}
\title{ How long does it take to give food in nursing home}
\description{
  A datasset of a students "evidence based nursing project (ebnp)"
  with the question how long it takes to give food to patients in a nursing home
  
}
\usage{data(Essen.Zeit)}
\format{
  A data frame with 63 observations on the following 22 variables.
  \describe{
    \item{\code{PATID}}{patient id}
    \item{\code{ALTER}}{patient's age}
    \item{\code{SEX}}{patient's sex}
    \item{\code{STUFE}}{patient's "Pflegestufe"}
    \item{\code{MAHLZEIT}}{patient's meal (supper, breakfast, lunch)}
    \item{\code{ORT}}{place}
    \item{\code{SPEILAGE}}{describing, where the meal was parked}
    \item{\code{PATLAGE}}{describing where the patient was sitting or lying}
    \item{\code{PATPOSIT}}{patient's heading section}
    \item{\code{NAHRART}}{describing wether the meal was reduced to small pieces}
    \item{\code{HILF1}}{help-medium 1}
    \item{\code{HILF2}}{help-medium 2}
    \item{\code{HILF3}}{help-medium 3}
    \item{\code{HILF1HOW}}{quantity of help-mediums}
    \item{\code{TRUNK1}}{drink}
    \item{\code{TRUNK1ML}}{quantity of drinking (in ml)}
    \item{\code{TRUNK2}}{drink 2}
    \item{\code{TRUNK2ML}}{quantity of drinking (in ml)}
    \item{\code{PERSON}}{person giving food to patient}
    \item{\code{ZEIT}}{time, how long it took to give food}
    \item{\code{BREAK}}{describing wether there was a break}
    \item{\code{PORTION}}{describing how much food was given}
    }
}

\keyword{datasets}
