\name{Mikrolagerung}
\alias{Mikrolagerung}
\docType{data}
\title{ pressure at sacral-bone: 30 degree vs. micro-positioning}
\description{
  Dataset of a german students "evidence based nursing project (ebnp)" with the question
  what pressure at the sacral-bone-area can be measured using 30 degree and micro-positioning
}
\usage{data(Mikrolagerung)}
\format{
  A data frame with 98 observations on the following 21 variables.
  \describe{
    \item{\code{PROBAND}}{ID}
    \item{\code{ALTER}}{age}
    \item{\code{GROESSE}}{height}
    \item{\code{GEWICHT}}{weigght}
    \item{\code{BMI}}{Body-Mass-Index}
    \item{\code{HUEFTE}}{Hip circumference}
    \item{\code{SEX}}{proband's sex}
    \item{\code{RETEST}}{test or retest}
    \item{\code{KOPFTEIL}}{position of head section}
    \item{\code{RE30}}{pressure at sacral-bone-area while lying 30 degree to the right side}
    \item{\code{SCHU_RE}}{true angel assessed at right shoulder during 30 degree positioning to the right side}
    \item{\code{BECK_RE}}{true angel assessed at right pelvis during 30 degree positioning to the right side}
    \item{\code{LI30}}{pressure at sacral-bone-area while lying 30 degree to the left side}
    \item{\code{SCHU_LI}}{true angel assessed at right shoulder during 30 degree positioning to the left side}
    \item{\code{BECK_LI}}{true angel assessed at right pelvis during 30 degree positioning to the left side}
    \item{\code{MIKRO1A}}{pressure at sacral-bone-area while micro-positioning at right shoulder}
    \item{\code{MIKRO1B}}{pressure at sacral-bone-area while micro-positioning at right pelvis}
    \item{\code{MIKRO1C}}{pressure at sacral-bone-area while micro-positioning at left shoulder}
    \item{\code{MIKRO1D}}{pressure at sacral-bone-area while micro-positioning at left pelvis}
    \item{\code{MIKRO2A}}{pressure at sacral-bone-area while micro-positioning at right pelvis and shoulder}
    \item{\code{MIKRO2B}}{pressure at sacral-bone-area while micro-positioning at left pelvis and shoulder}
  }
}
\details{
While lying in bed, pressure is the main reason why patients get a pressure-ulcer. 
The aim of this EBNP was to get knowledge about how pressure differs during various positionings.
The probands where positioned to the left and to the right side by 30 degrees while lying in a bed. Each time, the pressure at the 
sacral-bone-area was measured, as at this area a pressure-ulcer arises very often.
After that, pressure was assessed while probands where positioned by micro-positioning at left and right shoulder and pelvis.
}

\keyword{datasets}
