\name{wikibooks-package}
\alias{wikibooks-package}
\alias{wikibooks}
\docType{package}
\title{Functions and datasets of the german WikiBook "GNU R"}
\description{
	The german Wikibook "GNU R" introduces R to new users.
	This package is a collection of functions and datas used in the german WikiBook "GNU R"
}
\details{
\tabular{ll}{
Package: \tab wikibooks\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2007-06-04\cr
License: \tab GPL version 2 or newer \cr
}

}
\author{
Joerg Schlarmann (aka Produnis)

Maintainer: Joerg Schlarmann (aka Produnis) <schlarmann@produnis.de>
}
\references{
\url{http://de.wikibooks.org/wiki/GNU_R}
}
\keyword{ package }

